+++ .gitignore
@@ -1,0 +1,5 @@
+syntax: glob
+
+*.pyc
+build/*
+.idea
