--- .hgignore
@@ -1,4 +1,0 @@
-syntax: glob
-
-*.pyc
-build/*
