--- [test].txt
@@ -1,2 +1,0 @@
-test
-test
