+++ setup.py
@@ -1,10 +1,10 @@
 # setup.py, config file for distutils
 
-import __init__
-
-from distutils.core import setup
-from distutils.command.install_data import install_data
 import os
+from distutils.command.install_data import install_data
+from distutils.core import setup
+
+import unrar2
 
 
 class smart_install_data(install_data):
@@ -24,14 +24,14 @@ for dirpath, dirnames, filenames in os.walk(r'.'):
             pass
     for filename in [fn for fn in filenames if os.path.splitext(fn)[-1].lower() in ('.pyc', '.pyo', '.scc')]:
         filenames.remove(filename)
-    parts = ['UnRAR2']+dirpath.split(os.sep)[1:]
+    parts = ['unrar2']+dirpath.split(os.sep)[1:]
     
     data_files.append((os.path.join(*parts), [os.path.join(dirpath, fn) for fn in filenames]))
 
 setup(name='pyUnRAR2',
-      version=__init__.__version__,
+      version=unrar2.__version__,
       description='Improved Python wrapper around the free UnRAR.dll',
-      long_description=__init__.__doc__.strip(),
+      long_description=unrar2.__doc__.strip(),
       author='Konstantin Yegupov',
       author_email='yk4ever@gmail.com',
       url='http://code.google.com/py-unrar2',
@@ -47,8 +47,8 @@ setup(name='pyUnRAR2',
                    'Topic :: Software Development :: Libraries :: Python Modules',
                    'Topic :: System :: Archiving :: Compression',
                   ],
-      packages=['UnRAR2'],
-      package_dir={'UnRAR2' : ''},
+      packages=['unrar2'],
+      package_dir={'unrar2' : ''},
       data_files=data_files,
-      cmdclass = {'install_data': smart_install_data},
-     )
+      cmdclass = {'install_data': smart_install_data}, requires=['six']
+      )
