+++ tests/[test].txt
@@ -1,0 +1,2 @@
+test
+test
