+++ tests/test_protected_files.rar
